Folder support Part 1

+++ BookPlayer/Library/DataManagement/DataManager+CoreData.swift
@@ -15,28 +15,28 @@ extension DataManager {
     }
 
     /**
-     Creates a book for each URL and adds it to the specified playlist. If no playlist is specified, it will be added to the library.
+     Creates a book for each URL and adds it to the specified folder. If no folder is specified, it will be added to the library.
 
-     A book can't be in two places at once, so if it already existed, it will be removed from the original playlist or library, and it will be added to the new one.
+     A book can't be in two places at once, so if it already existed, it will be removed from the original folder or library, and it will be added to the new one.
 
      - Parameter files: `Book`s will be created for each element in this array
-     - Parameter playlist: `Playlist` to which the created `Book` will be added
-     - Parameter library: `Library` to which the created `Book` will be added if the parameter `playlist` is nil
+     - Parameter folder: `Folder` to which the created `Book` will be added
+     - Parameter library: `Library` to which the created `Book` will be added if the parameter `folder` is nil
      - Parameter completion: Closure fired after processing all the urls.
      */
-    class func insertBooks(from files: [FileItem], into playlist: Playlist?, or library: Library, completion: @escaping () -> Void) {
+    class func insertBooks(from files: [FileItem], into folder: Folder?, or library: Library, completion: @escaping () -> Void) {
         let context = self.getContext()
 
         for file in files.sorted(by: {$0.originalUrl.fileName < $1.originalUrl.fileName}) {
             // TODO: do something about unprocessed URLs
             guard let url = file.processedUrl else { continue }
 
-            // Check if book exists in the library
-            guard let item = library.getItem(with: url) else {
+            // Check if item exists in the library
+            guard let item = library.getItem(with: url.lastPathComponent) else {
                 let book = Book(from: file, context: context)
 
-                if let playlist = playlist {
-                    playlist.addToBooks(book)
+                if let folder = folder {
+                    folder.addToItems(book)
                 } else {
                     library.addToItems(book)
                 }
@@ -44,27 +44,15 @@ extension DataManager {
                 continue
             }
 
-            guard let storedPlaylist = item as? Playlist,
-                let storedBook = storedPlaylist.getBook(with: url) else {
-                // swiftlint:disable force_cast
-                // Handle if item is a book
-                let storedBook = item as! Book
-
-                if let playlist = playlist {
-                    library.removeFromItems(storedBook)
-                    playlist.addToBooks(storedBook)
-                }
-
-                continue
+            if let parent = item.folder {
+                parent.removeFromItems(item)
             }
 
-            // Handle if book already exists in the library
-            storedPlaylist.removeFromBooks(storedBook)
-
-            if let playlist = playlist {
-                playlist.addToBooks(storedBook)
+            if let folder = folder {
+                library.removeFromItems(item)
+                folder.addToItems(item)
             } else {
-                library.addToItems(storedBook)
+                library.addToItems(item)
             }
         }
 
@@ -76,7 +64,7 @@ extension DataManager {
     }
 
     /**
-     Creates a book for each URL and adds it to the library. A book can't be in two places at once, so it will be removed if it already existed in a playlist.
+     Creates a book for each URL and adds it to the library. A book can't be in two places at once, so it will be removed if it already existed in a folder.
 
      - Parameter bookUrls: `Book`s will be created for each element in this array
      - Parameter library: `Library` to which the created `Book` will be added
@@ -87,54 +75,56 @@ extension DataManager {
     }
 
     /**
-     Creates a book for each URL and adds it to the specified playlist. A book can't be in two places at once, so it will be removed from the library if it already existed.
+     Creates a book for each URL and adds it to the specified folder. A book can't be in two places at once, so it will be removed from the library if it already existed.
 
      - Parameter bookUrls: `Book`s will be created for each element in this array
-     - Parameter playlist: `Playlist` to which the created `Book` will be added
+     - Parameter folder: `Folder` to which the created `Book` will be added
      - Parameter completion: Closure fired after processing all the urls.
      */
-    public class func insertBooks(from files: [FileItem], into playlist: Playlist, completion: @escaping () -> Void) {
-        self.insertBooks(from: files, into: playlist, or: playlist.library!, completion: completion)
+    public class func insertBooks(from files: [FileItem], into folder: Folder, completion: @escaping () -> Void) {
+        self.insertBooks(from: files, into: folder, or: folder.library!, completion: completion)
     }
 
-    public class func delete(_ items: [LibraryItem], mode: DeleteMode = .deep) {
+    public class func delete(_ items: [LibraryItem], library: Library, mode: DeleteMode = .deep) {
         for item in items {
-            guard let playlist = item as? Playlist else {
+            guard let folder = item as? Folder else {
                 // swiftlint:disable force_cast
-                self.delete(item as! Book, mode: mode)
+                self.delete(item as! Book, library: library, mode: mode)
                 continue
             }
 
-            self.delete(playlist, mode: mode)
+            self.delete(folder, library: library, mode: mode)
         }
     }
 
-    public class func delete(_ playlist: Playlist, mode: DeleteMode = .deep) {
-        guard let library = playlist.library else { return }
+    public class func delete(_ folder: Folder, library: Library, mode: DeleteMode = .deep) {
 
         if mode == .shallow,
-            let orderedSet = playlist.books {
-            library.addToItems(orderedSet)
+            let orderedSet = folder.items {
+            if let parent = folder.folder {
+                parent.addToItems(orderedSet)
+            } else {
+                library.addToItems(orderedSet)
+            }
         }
 
         // swiftlint:disable force_cast
-        for book in playlist.books?.array as! [Book] {
+        for item in folder.items?.array as! [LibraryItem] {
             guard mode == .deep else { continue }
-            self.delete(book, mode: .deep)
+            self.delete(item, library: library, mode: .deep)
         }
 
-        library.removeFromItems(playlist)
+        library.removeFromItems(folder)
 
-        self.delete(playlist)
+        self.delete(folder)
     }
 
-    public class func delete(_ book: Book, mode: DeleteMode) {
+    public class func delete(_ item: LibraryItem, library: Library, mode: DeleteMode) {
         guard mode == .deep else {
-            if let playlist = book.playlist,
-                let library = playlist.library {
-                library.addToItems(book)
+            if let folder = item.folder {
+                library.addToItems(item)
 
-                playlist.removeFromBooks(book)
+                folder.removeFromItems(item)
 
                 self.saveContext()
             }
@@ -142,21 +132,23 @@ extension DataManager {
             return
         }
 
-        if book == PlayerManager.shared.currentBook {
-            NotificationCenter.default.post(name: .bookDelete,
-                                            object: nil,
-                                            userInfo: ["book": book])
-            PlayerManager.shared.stop()
-        }
+        if let book = item as? Book {
+            if book == PlayerManager.shared.currentBook {
+                NotificationCenter.default.post(name: .bookDelete,
+                                                object: nil,
+                                                userInfo: ["book": book])
+                PlayerManager.shared.stop()
+            }
 
-        let fileURL = book.fileURL
+            let fileURL = book.fileURL
 
-        DispatchQueue.global().async {
-            if let fileURL = fileURL {
-                try? FileManager.default.removeItem(at: fileURL)
+            DispatchQueue.global().async {
+                if let fileURL = fileURL {
+                    try? FileManager.default.removeItem(at: fileURL)
+                }
             }
         }
 
-        self.delete(book)
+        self.delete(item)
     }
 }
